removing error logging used during backend development, renaming file system
