--- backend/config/allowedOrigins.js
@@ -1,5 +1,0 @@
-const allowedOrigins = [
-    'http://localhost:3000',
-    'http://www.helphiveservices.com',
-    'http://helphiveservices.com'
-]
