--- backend/config/corsOptions.js
@@ -1,15 +1,0 @@
-const allowedOrigins = require('./allowedOrigins')
-
-const corsOptions = {
-    origin: (origin, callback) => {
-        if (allowedOrigins.indexOf(origin) !== -1 || !origin) {
-            callback(null, true)
-        } else {
-            callback(new Error('Not allowed by CORS'))
-        }
-    },
-    credentials: true,
-    optionsSuccessStatus: 200
-}
-
-module.exports = corsOptions
