+++ server/middleware/loginLimiter.js
@@ -1,5 +1,4 @@
 const rateLimit = require('express-rate-limit')
-const { logEvents } = require('./logger')
 
 const loginLimiter = rateLimit({
     windowMs: 60 * 1000, // 1 minute
@@ -7,7 +6,6 @@ const loginLimiter = rateLimit({
     message:
         { message: 'Too many login attempts from this IP, please try again after a 60 second pause' },
     handler: (req, res, next, options) => {
-        logEvents(`Too Many Requests: ${options.message.message}\t${req.method}\t${req.url}\t${req.headers.origin}`, 'errLog.log')
         res.status(options.statusCode).send(options.message)
     },
     standardHeaders: true, // Return rate limit info in the `RateLimit-*` headers
