+++ server/routes/authRoutes.js
