+++ server/server.js
@@ -2,12 +2,8 @@ require('dotenv').config()
 const express = require('express')
 const app = express()
 const path = require('path')
-const {logger, logEvents} = require('./middleware/logger')
-const errorHandler = require('./middleware/errorHandler')
 const cookieParser = require('cookie-parser')
-const cors = require('cors')
-const corsOptions = require('./config/corsOptions')
-const connectDB = require('./config/dbConnect')
+const connectDB = require('./config/connection.js')
 const mongoose = require('mongoose')
 const PORT = process.env.PORT || 3500
 
@@ -15,18 +11,14 @@ console.log(process.env.NODE_ENV)
 
 connectDB()
 
-app.use(logger) //custom middleware
-
-app.use(cors(corsOptions)) //third-party middleware
-
 app.use(express.json())
 
 app.use(cookieParser())
 
 app.use('/', express.static(path.join(__dirname, '/public'))) //built in middleware
 
-app.use('/', require('./routes/root'))
-app.use('/auth', require('./routes/authRoutes'))
+app.use('/', require('./routes/root.js'))
+app.use('/auth', require('./routes/authRoutes.js'))
 app.use('/users', require('./routes/userRoutes.js'))
 app.use('/notes', require('./routes/noteRoutes.js'))
 
@@ -41,8 +33,6 @@ app.all('*', (req, res) => {
     }
 })
 
-app.use(errorHandler)
-
 mongoose.connection.once('open', () => {
     console.log('Connected to MongoDB')
     app.listen(PORT, () => console.log(`Server running on port ${PORT}`))
@@ -50,5 +40,4 @@ mongoose.connection.once('open', () => {
 
 mongoose.connection.on('error', err => {
     console.log(err)
-    logEvents(`${err.no}: ${err.code}\t${err.syscall}\t${err.hostname}`, 'mongoErrLog.log')
 })
